services/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
httpx==0.27.0

opentelemetry-sdk==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-httpx==0.46b0
opentelemetry-instrumentation-starlette==0.46b0
pydantic==2.8.2
pyyaml==6.0.2
pytest==8.2.0